backend/requirements.txt
amqp==5.3.1
asgiref==3.9.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
diff-match-patch==20241021
Django==5.2.4
django-anymail==13.0
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-environ==0.12.0
django-import-export==4.3.8
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
humanize==4.12.3
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==12.3.0
tablib==3.8.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
